projects/rag/Enterprise-RAG/requirements.txt
# ============================================================
# Enterprise-RAG: Production-Grade RAG System Requirements
# ============================================================
# Python 3.11+ required
# All versions pinned for reproducibility

# =====================
# Core RAG Framework
# =====================
llama-index==0.10.55
llama-index-core==0.10.55
llama-index-llms-openai==0.1.25
llama-index-llms-anthropic==0.1.15
llama-index-vector-stores-chroma==0.1.10
llama-index-embeddings-huggingface==0.2.3
llama-index-postprocessor-cohere-rerank==0.1.6
llama-index-readers-file==0.1.26

# =====================
# Vector Databases
# =====================
chromadb==0.5.5
qdrant-client==1.11.3

# =====================
# Embeddings & Reranking
# =====================
sentence-transformers==2.6.1
transformers==4.40.1
torch==2.3.0
rank-bm25==0.2.2

# =====================
# Document Processing
# =====================
pypdf==4.2.0
python-docx==1.1.2
python-multipart==0.0.9
Pillow==10.3.0
python-magic==0.4.27

# =====================
# API Framework
# =====================
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.1
pydantic-settings==2.5.2
httpx==0.27.0
slowapi==0.1.9

# =====================
# Evaluation
# =====================
ragas==0.1.14
langchain==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.25
langchain-community==0.2.14
datasets==2.20.0

# =====================
# UI & Demo
# =====================
streamlit==1.39.0
altair==5.4.1

# =====================
# Utilities
# =====================
python-dotenv==1.0.1
click==8.1.7
rich==13.7.1
tqdm==4.66.4
aiofiles==23.2.1

# =====================
# Testing & Quality
# =====================
pytest==7.4.4
pytest-asyncio==0.23.7
pytest-cov==5.0.0
pytest-mock==3.14.0
black==24.8.0
ruff==0.6.4
mypy==1.11.1
pre-commit==3.8.0

# =====================
# Deployment
# =====================
docker==7.1.0
gunicorn==22.0.0
